Make basic use of MPI in order to verify MPI compilation and execution
In order to verify that SMIOL, as well as the SMIOL runner programs, will
successfully compile and link with MPI, add calls to MPI_Init and MPI_Finalize
in both the C and Fortran main programs.

This commit also adjusts the Makefile so that CC_PARALLEL and FC_PARALLEL
are used for compilation and linking.