Initial implementation of SMIOL_init
The SMIOL_init routine takes two arguments. The first is an MPI communicator
that is to be used by SMIOL, and the second is a pointer to a SMIOL_context
pointer. If the SMIOL_init routine completes successfully, the SMIOL_context
pointer will point to a valid SMIOL_context structure.

The SMIOL_context structure contains the size of the MPI communicator as well
as each task's rank in the communicator. Additionally, the structure contains
an MPI_Fint handle to the MPI communicator rather than an MPI_Comm type.
The choice of an MPI_Fint over an MPI_Comm was made in order to simplify
the definition of an interoperable SMIOL_context type in Fortran, where
an MPI_Fint is supposed to correspond to the default Fortran integer type.

At present, calling SMIOL_init will result in a memory leak, since there is
no method for freeing the SMIOL_context structure.