Add logic to Makefile for including and linking to the parallel-netCDF library
When the PNETCDF environment variable is define, the top-level Makefile will
define the CPPINCLUDES variable as "-DSMIOL_PNETCDF -I${PNETCDF}/include" and
the LIBS variable as "-L${PNETCDF}/lib -lpnetcdf". The CPPINCLUDES variable
is passed to the Makefile in the src/ subdirectory, and the LIBS variable
is used when linking the "runner" executables.