Initial implementation of SMIOL_define_dim
SMIOL_define_dim takes as arguments a file handle, a dimension name, and
a dimension size. The dimension size was chosen to be of type int64_t rather
than uint64_t because the Fortran ISO_C_BINDING module does not define
an interoperable type for a C uint64_t, probably because Fortran does not have
unsigned integer types. However, this choice allows for the use of negative
dimension sizes to indicate record or unlimited dimensions.