Provide parallel-netCDF implementation of SMIOL_inquire_dim
With the ability to inquire about the size of dimensions in a file that has been
created by SMIOL_open_file, the unit tests for SMIOL_inquire_dim can now compare
the expected size of dimensions with the sizes used when defining
the dimensions.

If SMIOL is compiled without the parallel-netCDF library, SMIOL_inquire_dim is
expected to always succeed and return a zero value for the dimension size.
Accordingly, there are some pre-processing directives in unit tests to verify
this behavior when SMIOL_PNETCDF is not defined.