Add tests for parallel-netCDF library errors in opening and closing files
The first of these tests attempts to create a file in the root of
the filesystem, which we assume will fail.

The second test attempts to close a file that was never opened. This test
requires hacking a SMIOL_file instance to point to a valid SMIOL_context
without containing a valid file handle.

These new tests are only active if SMIOL was compiled with support for
the parallel-netCDF library; if not, we have no way of actually trying
to open or close a file, and therefore cannot really test parallel-netCDF
error codes.