Initial implementation of SMIOL_file_sync
This commit adds the initial implementation of SMIOL_file_sync and
SMIOLf_file_sync, which sends valid arguments to SMIOL_file_sync. Currently,
the SMIOL_file_sync routine does not do any sort of 'syncing' (i.e. moving
in-memory data to the file system), it only checks to see if the given file
argument is not NULL and returns appropriate error codes if it is.

A future commit, or commits, can add specific libraries for writing, like
parallel-netCDF, which will actually call routines that will move data from
in-memory to the file system.