Implementation of parallel-netCDF to SMIOL_file_sync
If SMIOL is compiled with parallel-NetCDF, then ncmpi_sync will be called to
flush the system cache to the file system. If SMIOL is not compiled with
PnetCDF, then SMIOL_file_sync returns SMIOL_SUCCESS, upon a valid file handle.

If SMIOL is compiled with parallel-netcdf, then the SMIOL_file_sync function
will check the file handle to determine if the file is in 'define' mode or not.
If it is in 'define' mode, then it will call `ncmpi_enddef`, update the file
handle, and then call `ncmpi_sync`.

In the future, other libraries (or other methods of I/O) can add their methods
of flushing and syncing to SMIOL_file_sync.