Add pNetCDF state member to SMIOL_file
A new member has been added to the SMIOL_file struct and SMIOLf_file derived
type: state. The state member can be used to keep track of a file's
'mode' or 'status'.

At present, SMIOL only uses the state member when SMIOL is compiled with
PnetCDF, if its not compiled with PnetCDF, the state is unused.

If compiled with PnetCDF then SMIOL will use the state member to keep
track of a PnetCDF file's 'mode', i.e. whether it is in define mode or data
mode.