Initial implementation of arguments for SMIOL(f)_define_var
As of this commit, SMIOL_define_var does not actually define variables in
a file. Rather, the purpose of the changes in this commit is to verify that
arrays of strings can be correctly passed through the Fortran interface to
the C implementation of SMIOL_define_var.

Note: At present, this code does not build with the Cray compilers. More
clean-up and comments are needed.