Add new unit tests for inquiring about dimensions in existing files
Previously, there were only unit tests to ensure that a dimension could be
defined in a new file and queried from that file, with no intervening close
and re-open of the file. These tests would not have caught a situation in
which the file library was caching dimension values in memory but not actually
writing them to disk.

The new unit tests added in this commit read back dimensions after a file
has been closed and re-opened.