Switch file to define mode in SMIOL_define_dim
Previously, if a file was in NetCDF data mode and `SMIOL_define_dim` was called
PnetCDF would return an error when trying to define a new dimension. This
commit adds logic to switch a file out of data mode and into define mode in
before the call to `ncmpi_def_dim`.