Initial implementation of SMIOL(f)_inquire_var
This commit provides an initial implementation of the routines for
inquiring about the properties of a variable in a file.

A new variable type code has been added to smiol_codes.inc to
indicate an unknown variable type.

As of this commit, there are no unit tests for the SMIOL(f)_inquire_var
routines.