Have SMIOLf_create_decomp validate array size
This commit contains two changes to the SMIOLf_create_decomp routine. First, it
checks to see if the reported number of n_compute_elements and n_io_elements
matches the size (retrieved from Fortran's size statement) of compute_elements
and io_elementes, respectively.

It then adds a check to avoid calling c_loc on a zero-sized array by checking
to see if the size of the array is greater than zero. If the array is
zero-sized, then a c_null_pointer will be used for the c_compute_elements or
c_io_elements.