Update the SMIOL(f)_create_decomp interfaces
In preparation for an initial implementation of the SMIOL_create_decomp
routine, the SMIOL(f)_create_decomp interfaces have been updated to include
a SMIOL_context argument, and the order of the remaining arguments has been
changed so that the number of elements precedes the list of elements for
both compute and I/O elements.

Additionally, uses of int64 or c_int64_t have been replaced with SMIOL_Offset
(in C) or SMIOL_offset_kind (in Fortran).