Update SMIOL_decomp structure with send/receive lists for compute and I/O tasks
The SMIOL_create_decomp routine is expected to eventually return lists
describing the local elements that each MPI task needs to send and receive.
In anticipation of this implementation, this commit updates the SMIOL_decomp
structure to contain appropriate lists.

The example C code, below, illustrates the expected structure of a list:

        j = 0;
        n_neighbors = comp_list[j++];
        for (i = 0; i < n_neighbors; i++) {
                send_to = comp_list[j++];
                n_send = comp_list[j++];
                printf("----- send %i elements to %i -----\n", n_send, send_to);
                for (k = 0; k < n_send; k++) {
                        printf("  %i\n", comp_list[j+k]);
                }
                j += n_send;
        }

At present, the SMIOL_create_decomp routine has been minimally updated to
work with the new members of the SMIOL_decomp structure.