Update SMIOL_create_decomp to return an integer status code
To be more consistent with other SMIOL library calls, this commit changes the
return type of SMIOL_create_decomp to return an integer status which is used to
returned SMIOL status codes.

It also adds check to see if either compute_elements or io_elements are NULL
and either n_compute_elements or n_io_elements are not 0, respectively. If
compute_elements or io_elements is NULL and n_compute_elements or n_io_elements
is not 0, then SMIOL_INVALID_ARGUMENT is returned.

It also adds a check to see if the passed in context is NULL.