Fix use of uninitialized memory in SMIOLf_inquire_var
Valgrind's memcheck tool identified a use of uninitialized memory that
occurs when copying dimension names to the dimnames output argument
in SMIOLf_inquire_var:

==6976== Conditional jump or move depends on uninitialised value(s)
==6976==    at 0x417211: __smiolf_MOD_smiolf_inquire_var (smiolf.F90:707)
==6976==    by 0x40D0B7: test_variables.3801 (smiol_runner.F90:1279)
==6976==    by 0x404963: MAIN__ (smiol_runner.F90:82)
==6976==    by 0x414A7F: main (smiol_runner.F90:6)

This commit fixes this issue by always requesting the actual number of
dimensions for a variable if the dimnames argument is present, and only
copying out as many dimension names as the variable has dimensions.

This commit also modifies the SMIOLf_inquire_var routine so that it returns
after an unsuccessful call to SMOIL_inquire_var before trying to copy
values to output arguments, as there is no point in trying to set output
arguments based on potentially uninitialized memory.