Add SMIOL utility routines for searching and sorting
This commit adds a new code file, smiol_utils.c, along with a header file to
provide routines for searching and sorting. At present, there is one routine
for sorting arrays of triplets of SMIOL_Offset values and one routine for
searching sorted arrays of triplets of SMIOL_Offset values.

TO DO: The files with utility functions will ultimately need a way to obtain
       type definitions like SMIOL_Offset. In this commit, the utility functions
       simply use int64_t.