Add new unit tests for SMIOL(f)_create_decomp with 2 MPI tasks
In order to test that the SMIOL_create_decomp routine is actually working out
correct comp_list and io_list arrays, this commit adds new unit tests for
SMIOL(f)_create_decomp that are only active when the SMIOL unit tests are run
with exactly two MPI tasks.

At present, the simplest way to verify the results for known patterns of compute
elements and I/O elements is to work out manually what the comp_list and io_list
arrays are expected to contain, then to compare these known-correct arrays with
those that are returned by the SMIOL_create_decomp routine.