Add inital placeholder arguments to SMIOL(f)_put_var
This commit also changes the `SMIOLf_put_var` function into a generic interface
and creates four a four new `SMIOLf_put_var_(type)` routines. Where type is
`char`, `int`, `real32` or `real64`.

Because none of the Fortran dummy arguments in the new SMIOLf_put_var_(type)
routines, this commit creates a lot of warnings which will be resolved in the
next commit.