Initial implementation of SMIOL(f)_put_var
This commit adds the initial implementation of SMIOL(f)_put_var. In order for
the `buf` in smiolf.F90's SMIOLf_put_var_(type) routines to be interoperable
with C for 2003 Fortran, a new argument was added (`bufsize`) to make `buf` an
explicit sized array.

This commit also removes all of the warnings created in 3aa72d1.