Add unit tests for transfer_field utility function
The unit tests in this commit make use of the same decomp/io patterns that
were (in principle) verified to be correct in the test_decomp function. Using
these known-good decomps, the unit tests:
1) Transfer a field with a simple pattern of values from compute to I/O tasks
2) let the I/O tasks modify the field
3) let the compute tasks zero-out their original field
4) transfer the modified field from I/O to compute tasks
5) and return a success code if the field on the compute tasks contains
   the original pattern with the expected modification.

The unit tests transfer fields of several different types: char, int, float, and
double. A cpp macro is used to define the test function for each of these types.