Avoid memory leaks in SMIOL_create_decomp after malloc failure
When calls to malloc return a NULL pointer in SMIOL_create_decomp, any memory
that has been malloc'ed up to the point of failure in the routine is now
free'd to avoid a memory leak before SMIOL_MALLOC_FAILURE is returned. These
changes also ensure that the decomp pointer is NULL whenever SMIOL_create_decomp
returns with an exit code other than SMIOL_SUCCESS.

Note that for the call to malloc inside the loop that iterates over all MPI
ranks, we do not yet try to catch any error in malloc'ing the new incoming
buffer in each iteration because there is not yet a way to broadcast that
error to all other MPI ranks and to exit gracefully.