PnetCDF implementation of SMIOL_put_var
This commit contains the PnetCDF implementation in SMIOL_put_var, which enables
SMIOL to write variable data to disk.

Because there isn't a way to allocate the start, and count of the SMIOLf_decomp
structure yet, this commit disables the Fortran SMIOLf_put_var test, which
causes a segmentation fault.