Update SMIOL(f)_inquire_dim to inquire about unlimited dims
This commit adds a new argument to SMIOL_inquire_dim and SMIOLf_inquire_dim,
is_unlimited, which allows inquires on whether the requested dimension
is the unlimited dimension or not. This commit also makes the dimsize
argument optional, which makes inquire_dim similar to the functionality of
inquire_var. This commit updates all calls to the SMIOL_inquire_dim function in
the unit tests to contain the new argument.