Define two new return codes: SMIOL_WRONG_ARG_TYPE and SMIOL_INSUFFICIENT_ARG
The SMIOL_WRONG_ARG_TYPE code may be returned in situations where, e.g.,
the caller has inquired about an integer attribute using a real-valued variable.

The SMIOL_INSUFFICIENT_ARG code may be returned when it can be determined that
the actual argument to a function is too small to contain the value in a file.