Add SMIOLf_put/get_var unit tests
Currently, the test for SMIOL_get_var for characters does not pass. This is
because the result from ncmpi_get_vara is not transfered back into the
character buf provided by the user. In other interfaces, c_loc is used on the
buf so that the read of SMIOL_get_var is read into buf, but for characters, the
string read from SMIOL_get_var is read into an intermediate variable and not
transfered to the original buffer.