Add frame and set/get functions to SMIOL C
This commit adds an additional field to the SMIOL_file struct, frame, which
will keep track of current frame for the unlimited dimension of an open file.
When a file is opened or created, frame is set to 0.

This commit also adds two new functions, SMIOL_set_frame and SMIOL_get_frame
which can set the frame of a file or get the current frame of a file,
respectively.