Ensure that vartype and ndims arguments are always defined by SMIOL_inquire_var
In case SMIOL is not compiled with support for the parallel-netCDF library,
the output arguments vartype and ndims were previously never given values by
the SMIOL_inquire_var routine. This commit adds code to initialize vartype
to SMIOL_UNKNOWN_VAR_TYPE and ndims to 0; dimnames can't reasonably be
initialized, since a caller might expect that ndims is zero and therefore
not allocate any strings for dimension names.