Checks in SMIOL_put/get_var unit tests that MPI task count evenly divides 120
Some of the unit tests for SMIOL_put/get_var decompose a dimension (nCells) of
size 120, and their logic requires that the number of MPI tasks evenly divides
nCells. This commit adds simple logic to check whether 120 is evenly divided,
and only executes unit tests that depend on the decomposition of nCells if
this is the case.

This commit also adds -lm to the build command for smiol_runner_c, since
the new check makes use of the floor and ceil functions.