Avoid zero-sized c_dimnames argument to C_LOC in SMIOLf_define_var
In SMIOLf_define_var, the definition of variables with zero dimensions
(i.e., scalar variables) led to a situation in which the argument of
the C_LOC intrinsic was a zero-sized array. Specifically, this argument
was the c_dimnames array.

According to the Fortran 2003 standard, the argument of C_LOC shall be
"...an allocated allocatable variable that has the TARGET attribute and is not
an array of zero size..."

To avoid this problem, an actual argument, c_dimnames_ptr, is declared in
SMIOLf_define_var, and is set to either C_LOC(c_dimnames) or C_NULL_PTR,
depending on whether the variable being defined has more than zero dimensions.