Avoid zero-sized compute_elements argument to C_LOC in SMIOLf_create_decomp
In SMIOLf_create_decomp, MPI ranks with no compute elements pass a zero-sized
array as the compute_elements argument, and this led to a call to C_LOC with
a zero-sized array.

According to the Fortran 2003 standard, the argument of C_LOC shall be
"...an allocated allocatable variable that has the TARGET attribute and is not
an array of zero size..."

Now, if the size of the compute_elements array is zero, the C pointer to
compute_elements, c_compute_elements, is set to C_NULL_PTR; otherwise, this
pointer is set using C_LOC(compute_elements) as before.