Fix computation of io_count when comm_rank >= io_stride * num_io_tasks
The get_io_elements function previously assumed that the product of io_stride
and num_io_tasks would be at least as large as comm_size. For MPI ranks that
were a multiple of io_stride and greater than or equal to
(io_stride * num_io_tasks), the get_io_elements utility function incorrectly
returned non-zero io_start and io_count values.

The simple fix to this issue is to return from get_io_elements as soon as
io_rank has been computed as (comm_rank / io_stride) and been found to be equal
to or to exceed the number of I/O tasks (equality necessitates a return, since
io_rank is zero-based).