Implement SMIOLf_put_var and SMIOLf_get_var generic interfaces
This commit adds implementations of the SMIOLf_put_var and SMIOLf_get_var
generic interfaces for 0-d character variables, {1,2,3,4}-d int32 variables,
and {1,2,3,4,5}-d real32 and real64 variables. The specific routines were all
generated by a shell script, gen_put_get.sh, and reside in an include file,
smiolf_put_get_var.inc, that is included by the main SMIOLf module through
a pre-processor #include directive.

The shell script that was used to generate the routines is not run
automatically during the build process; accordingly, any manual edits to
the smiolf_put_get_var.inc file will be preserved until the script is explicitly
run again.

Also included in this commit are minimal updates to the Fortran unit test
program to enable it to compile with the new arguments to the SMIOLf_put_var
and SMIOLf_get_var routines.