Initialize char_buf in SMIOLf_get_var_0d_char to avoid returning garbage
The SMIOLf_get_var_0d_char routine internally allocates a character array,
char_buf, into which the requested variable is read; char_buf is then
transferred to the actual output argument, which is a Fortran string.

In cases where the actual argument supplied by the caller is longer than
the size of the variable read from a file, trailing bytes in char_buf would be
uninitialized, and would therefore result in trailing garbage being returned to
the caller even if the caller had initialized the actual argument before
the call to SMIOLf_get_var_0d_char.

This commit adds a loop to initialize char_buf with the contents of buf in
the SMIOLf_get_var_0d_char routine immediately after char_buf has been allocated,
so that any un-read bytes will be preserved with their value just before the call
to SMIOLf_get_var_0d_char.