Correctly compute element_size for non-decomposed variables in build_start_count
The build_start_count function previously only computed a correct element size
for decomposed variables; non-decomposed variables did not result in an element
size that was scaled according to the type of the variable. However, there is no
reason why this function cannot return a correct element size for non-decomposed
variables, and this commit updates the build_start_count function to do so, as
the element size may be useful in future in SMIOL_put_var and SMIOL_get_var.