Add aggregation_factor argument to SMIOL_create_decomp
The aggregation_factor argument is an optional argument to the Fortran
SMIOLf_create_decomp routine, and if not present, is assumed to be 0, which
indicates that SMIOL may choose whatever aggregation factor it decides is
appropriate.

All C and Fortran unit tests have been updated to pass a value of 1 to
the aggregation_factor argument of SMIOL_create_decomp, which effectively
requests no aggregation.

At present, the aggregation factor provided to SMIOL_create_decomp isn't
actually used.