Add unit tests for creating decompositions with aggregation
The degree of aggregation affects the decomposition, and previous tests of the
SMIOL_create_decomp routine only used an aggregation factor of 1 (i.e., no
aggregation). This commit adds two tests that require exactly four MPI ranks:
the first test uses an aggregation factor of 2 and an I/O stride of 1; and
the second test uses an aggregation factor of 3 and an I/O stride of 2.