Omit (through preprocessing) several C unit tests with invalid inputs
Since future development in the SMIOL_close_file and SMIOL_sync_file routines
will make use of the members of SMIOL_file structs, and since at present there
is no way to check on the validity of the memory associated with a SMIOL_file
struct, two C unit tests that passed invalid SMIOL_file structs (essentially
allocated memory with random contents) are now omitted.

If in future there were a way to verify the validity of a SMIOL_file struct,
validity could be checked in the SMIOL_close_file and SMIOL_sync_file functions
before making use of members of the SMIOL_file struct, and the two tests that
have been pre-processed out in this commit could be re-enabled.