Silence compilation warnings when SMIOL_PNETCDF is not defined
When compiling without the use of the Parallel-NetCDF library (i.e., when
the SMIOL_PNETCDF macro is not defined), the GNU compiler warned about
several instances in which the io_group_comm variable was set but not used:

  warning: variable ‘io_group_comm’ set but not used

The io_group_comm communicator is used in many places to communicate errors
from the underlying I/O library (Parallel-NetCDF) to non-I/O tasks, and if
no underlying I/O library is used, then there is no need to communicate.

Now, the io_group_comm variable is only declared and set if SMIOL_PNETCDF
is defined.