Fix bug when inquiring about dimension names with non-unit I/O stride
When the I/O stride is larger than 1, the SMIOL_inquire_var routine must
broadcast values read from the file by I/O tasks to non-I/O tasks. In the case
of dimension names, the SMIOL_inquire_var routine previously broadcast each
dimension name with a broadcast size of 'len' characters, where len is the
result of a call to strnlen. Since the strnlen function returns the length of a
string excluding the terminating null character, dimension names were broadcast
to non-I/O tasks without a terminating null character, resulting in incorrect
strings on non-I/O tasks. The simple fix for this is to increment the string
length, 'len', before using it as the size of the broadcast for dimension names.