Modify unit tests for SMIOL_inquire_var to catch bad dimnames on non-I/O task
This commit updates the unit tests for SMIOL(f)_inquire_var (as well as other
SMIOL variable routines) in the C and Fortran 'test_variables' functions so that
a non-unit I/O stride is used for file access. The purpose of this change is to
catch future errors in which returned values, e.g., for dimnames, are incorrect
on non-I/O tasks. With the unit-stride file access used in these tests prior to
this commit, all tasks read/wrote values to a file, and non-unit-stride
correctness was therefore not tested.